analyzer/requirements.txt
# Core dependencies
ultralytics==8.3.40
torch==2.9.1
torchvision==0.24.1
opencv-python==4.10.0.84
numpy==2.1.3
scipy==1.15.1

# Video processing
imageio==2.36.1
imageio-ffmpeg==0.5.1

# Kafka
kafka-python==2.2.2

# Database
psycopg2-binary==2.9.10
redis==5.2.0

# Object storage
minio==7.2.18

# Web framework (for potential API endpoints)
flask==3.1.0
requests==2.32.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
scikit-learn==1.5.2
pandas==2.2.3

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0